// lib/types.ts
export interface Transaction {
  id: string;
  type: "subscription" | "referral" | "refund";
  amount: number;
  mode: "platform" | "card";
  status: "completed" | "pending" | "failed";
  plan?: string;
  date: Date;
  cardLast4?: string;
  description: string;
  paymentMethod?: string;
}

export interface Invoice {
  id: string;
  number: string;
  status: "paid" | "open" | "void" | "uncollectible";
  amount_due: number;
  amount_paid: number;
  currency: string;
  created: string;
  hosted_invoice_url?: string;
  invoice_pdf?: string;
  subscription?: string;
  billing_reason?: string;
  discount?: any;
  description?: string;
}